# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse API
  description: Self-hosted uptime monitoring platform API.
  version: 1.0.0
  contact:
    name: Pulse
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
paths:
...
  /monitors:
    get:
      operationId: listMonitors
      summary: List all monitors
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of monitors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createMonitor
      summary: Create a new monitor
      tags: [Monitors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonitorRequest"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /monitors/{id}:
    get:
      operationId: getMonitor
      summary: Get a monitor by ID
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "200":
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putMonitor
      summary: Create or update a monitor (idempotent)
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutMonitorRequest"
      responses:
        "200":
          description: Monitor created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMonitor
      summary: Delete a monitor
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "204":
          description: Monitor deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /monitors/{id}/stats:
    get:
      operationId: getMonitorStats
      summary: Get uptime statistics, SSL info, and last error for a monitor
      tags: [Monitors]
      parameters:
        - $ref: "#/components/parameters/MonitorID"
      responses:
        "200":
          description: Monitor statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorStatsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API Token
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    MonitorID:
      name: id
      in: path
      required: true
      description: Monitor UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Monitor:
      type: object
      required:
        - id
        - name
        - type
        - target
        - interval_seconds
        - timeout_seconds
        - status
        - state
        - settings
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [http, tcp, udp, websocket]
        target:
          type: string
        interval_seconds:
          type: integer
        timeout_seconds:
          type: integer
        status:
          type: string
          enum: [active, paused]
        state:
          type: string
          enum: [up, down, unknown]
        last_checked_at:
          type: string
          format: date-time
          nullable: true
        next_check_at:
          type: string
          format: date-time
          nullable: true
        settings:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMonitorRequest:
      type: object
      required: [name, type, target]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [http, tcp, udp, websocket]
        target:
          type: string
        interval_seconds:
          type: integer
          default: 60
        timeout_seconds:
          type: integer
          default: 10
        status:
          type: string
          enum: [active, paused]
          default: active
        settings:
          type: object
    PutMonitorRequest:
      type: object
      required: [name, type, target]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [http, tcp, udp, websocket]
        target:
          type: string
        interval_seconds:
          type: integer
          default: 60
        timeout_seconds:
          type: integer
          default: 10
        status:
          type: string
          enum: [active, paused]
          default: active
        settings:
          type: object
    MonitorListResponse:
      type: object
      required: [data, total, page, limit, total_pages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Monitor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
...
    MonitorStatsResponse:
      type: object
      required: [monitor_id, uptime_24h, uptime_30d]
      properties:
        monitor_id:
          type: string
          format: uuid
        uptime_24h:
          $ref: "#/components/schemas/UptimeWindowStats"
        uptime_30d:
          $ref: "#/components/schemas/UptimeWindowStats"
        ssl:
          $ref: "#/components/schemas/SSLInfo"
        last_error:
          $ref: "#/components/schemas/LastErrorInfo"
    UptimeWindowStats:
      type: object
      required: [total_checks, up_checks, uptime_percent, avg_latency_ms]
      properties:
        total_checks:
          type: integer
        up_checks:
          type: integer
        uptime_percent:
          type: number
          format: float
        avg_latency_ms:
          type: integer
    SSLInfo:
      type: object
      required: [days_remaining, expires_at]
      properties:
        days_remaining:
          type: integer
          description: Days until certificate expires
        expires_at:
          type: string
          format: date
          description: Approximate certificate expiration date
    LastErrorInfo:
      type: object
      required: [error, checked_at]
      properties:
        error:
          type: string
          description: The error message from the last failed check
        checked_at:
          type: string
          format: date-time
          description: When the error was recorded